Sistema de roles terminado Fixed #127 Fixed #128

+++ resources/views/bombero/lista.blade.php
@@ -38,12 +38,18 @@
               <td class="text-center">{{$bombero->direccion}}</td>
               <td class="text-center">{{$bombero->telefono}}</td>
               <td class="text-center">{{$bombero->fecha_nacimiento}}</td>
-              <td class="text-center">
-                {{ Form::open(['route' => ['bombero.destroy', $bombero->id], 'method' => 'DELETE']) }}
-                    <button type="submit" class="btn glyphicon glyphicon-trash simulara"></button>
-                {{ Form::close() }}
-              </td>
-              <td class="text-center"><a class="glyphicon glyphicon-edit" href="{{ route('bombero.edit', $bombero->id) }}"></a></td>
+              @if (Auth::user()->admin)
+                <td class="text-center">
+                  {{ Form::open(['route' => ['bombero.destroy', $bombero->id], 'method' => 'DELETE']) }}
+                      <button type="submit" class="btn glyphicon glyphicon-trash simulara"></button>
+                  {{ Form::close() }}
+                </td>
+                <td class="text-center"><a class="glyphicon glyphicon-edit" href="{{ route('bombero.edit', $bombero->id) }}"></a></td>
+              @else
+                <td class="text-center" colspan="2">
+                  <button type="submit" class="btn glyphicon glyphicon-ban-circle ban" title="Sin permisos para eliminar/modificar"></button>
+                </td>
+              @endif
             </tr>
           @endforeach
           </tbody>
